Initial implementation of filters for Signature genesets.  this implementation added filter options to the GIU and allows users to filter by two options.  First,  absolute number of genes in the overlap between signature sets and EM sets.  (for example, only keep signature sets that have at least 5 genes in at least one of the EM genesets).  Second, percentage of overlap between signature sets and EM sets. (for examples, only keep signature sets where the overlap between the signature set and the EM set is at least X percentage of the EM geneset).
This feature will help in filtering out genesets from GMT files where one is not sure which of the genesets they are searching for but additional filters need to added to optimize different scenarios, i.e. for the drug - target genesets want to be able to highlight drugs that have targets in the given enrichment map but that are also specific (i.e. only has few targets)